+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -22,6 +22,8 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         SpacingAroundUnaryOperatorRule(),
         AnnotationSpacingRule(),
         UnnecessaryParenthesesBeforeTrailingLambdaRule(),
+        TypeParameterListSpacingRule(),
+        TypeArgumentListSpacingRule(),
         BlockCommentInitialStarAlignmentRule(),
         DiscouragedCommentLocationRule(),
         FunKeywordSpacingRule(),