+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/SpacingAroundOperatorsRule.kt
@@ -11,7 +11,6 @@ import com.pinterest.ktlint.core.ast.ElementType.EQEQ
 import com.pinterest.ktlint.core.ast.ElementType.EQEQEQ
 import com.pinterest.ktlint.core.ast.ElementType.EXCLEQ
 import com.pinterest.ktlint.core.ast.ElementType.EXCLEQEQEQ
-import com.pinterest.ktlint.core.ast.ElementType.FUN
 import com.pinterest.ktlint.core.ast.ElementType.GT
 import com.pinterest.ktlint.core.ast.ElementType.GTEQ
 import com.pinterest.ktlint.core.ast.ElementType.LT
@@ -20,17 +19,15 @@ import com.pinterest.ktlint.core.ast.ElementType.MINUS
 import com.pinterest.ktlint.core.ast.ElementType.MINUSEQ
 import com.pinterest.ktlint.core.ast.ElementType.MUL
 import com.pinterest.ktlint.core.ast.ElementType.MULTEQ
+import com.pinterest.ktlint.core.ast.ElementType.OPERATION_REFERENCE
 import com.pinterest.ktlint.core.ast.ElementType.OROR
 import com.pinterest.ktlint.core.ast.ElementType.PERC
 import com.pinterest.ktlint.core.ast.ElementType.PERCEQ
 import com.pinterest.ktlint.core.ast.ElementType.PLUS
 import com.pinterest.ktlint.core.ast.ElementType.PLUSEQ
-import com.pinterest.ktlint.core.ast.ElementType.TYPE_PARAMETER_LIST
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_ARGUMENT
-import com.pinterest.ktlint.core.ast.ElementType.WHITE_SPACE
 import com.pinterest.ktlint.core.ast.isPartOf
 import com.pinterest.ktlint.core.ast.nextLeaf
-import com.pinterest.ktlint.core.ast.parent
 import com.pinterest.ktlint.core.ast.prevLeaf
 import com.pinterest.ktlint.core.ast.upsertWhitespaceAfterMe
 import com.pinterest.ktlint.core.ast.upsertWhitespaceBeforeMe
@@ -40,8 +37,6 @@ import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafElement
 import org.jetbrains.kotlin.com.intellij.psi.tree.TokenSet
 import org.jetbrains.kotlin.psi.KtImportDirective
 import org.jetbrains.kotlin.psi.KtPrefixExpression
-import org.jetbrains.kotlin.psi.KtSuperExpression
-import org.jetbrains.kotlin.psi.KtTypeArgumentList
 
 class SpacingAroundOperatorsRule : Rule("op-spacing") {
 
@@ -58,40 +53,20 @@ class SpacingAroundOperatorsRule : Rule("op-spacing") {
         if (tokenSet.contains(node.elementType) &&
             node is LeafElement &&
             !node.isPartOf(KtPrefixExpression::class) && // not unary
-            !node.isPartOf(KtTypeArgumentList::class) && // C<T>
             !(node.elementType == MUL && node.treeParent.elementType == VALUE_ARGUMENT) && // fn(*array)
-            !node.isPartOf(KtImportDirective::class) && // import *
-            !node.isPartOf(KtSuperExpression::class) // super<T>
+            !node.isPartOf(KtImportDirective::class) // import *
         ) {
-            if (node.elementType == LT || node.elementType == GT) {
-                val list = node.parent(TYPE_PARAMETER_LIST)
-                if (list != null) {
-                    if (node.elementType == LT) {
-                        // ensure no space after < in <T>
-                        val nextLeaf = node.nextLeaf()
-                        if (nextLeaf?.elementType == WHITE_SPACE && !nextLeaf.textContains('\n')) {
-                            emit(node.startOffset + 1, "Unexpected spacing after \"${node.text}\"", true)
-                            if (autoCorrect) {
-                                nextLeaf.treeParent.removeChild(nextLeaf)
-                            }
-                        }
-                    } else {
-                        // ensure no space before > in <T>
-                        val prevLeaf = node.prevLeaf()
-                        if (prevLeaf?.elementType == WHITE_SPACE && !prevLeaf.textContains('\n')) {
-                            emit(prevLeaf.startOffset, "Unexpected spacing before \"${node.text}\"", true)
-                            if (autoCorrect) {
-                                prevLeaf.treeParent.removeChild(prevLeaf)
-                            }
-                        }
-                    }
-                    if (node.parent(TYPE_PARAMETER_LIST)?.treeParent?.elementType != FUN) {
-                        return // FIXME: only `fun <T>fn(): T {}` is checked
-                    }
-                }
+            if ((node.elementType == LT || node.elementType == GT || node.elementType == MUL) &&
+                node.treeParent.elementType != OPERATION_REFERENCE
+            ) {
+                // Do not format parameter types like:
+                //   <T> fun foo(...)
+                //   class Foo<T> { ... }
+                //   Foo<*>
+                return
             }
-            val spacingBefore = node.prevLeaf() is PsiWhiteSpace || node.elementType == GT
-            val spacingAfter = node.nextLeaf() is PsiWhiteSpace || node.elementType == LT
+            val spacingBefore = node.prevLeaf() is PsiWhiteSpace
+            val spacingAfter = node.nextLeaf() is PsiWhiteSpace
             when {
                 !spacingBefore && !spacingAfter -> {
                     emit(node.startOffset, "Missing spacing around \"${node.text}\"", true)