+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/SpacingAroundOperatorsRuleTest.kt
@@ -2,6 +2,8 @@ package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.test.diffFileFormat
 import com.pinterest.ktlint.test.diffFileLint
+import com.pinterest.ktlint.test.format
+import com.pinterest.ktlint.test.lint
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
@@ -20,5 +22,15 @@ class SpacingAroundOperatorsRuleTest {
                 "spec/op-spacing/format-expected.kt.spec"
             )
         ).isEmpty()
+    }
+
+    @Test
+    fun `Given a type argument list, containing a wildcard type then do not require space around the wildcard`() {
+        val code =
+            """
+            val foo = Map<PropertyType<*>>
+            """.trimIndent()
+        assertThat(SpacingAroundOperatorsRule().lint(code)).isEmpty()
+        assertThat(SpacingAroundOperatorsRule().format(code)).isEqualTo(code)
     }
 }